requirements.txt
agno
python-dotenv
openai
lancedb
streamlit
fastapi
uvicorn
pypdf